% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatherArray-package.R, R/FeatherArray.R
\docType{package}
\name{FeatherArray-package}
\alias{FeatherArray-package}
\alias{dim,FeatherArraySeed-method}
\alias{dimnames,FeatherArraySeed-method}
\title{FeatherArray: Using feather as a DelayedArray Backend}
\usage{
\S4method{dim}{FeatherArraySeed}(x)

\S4method{dimnames}{FeatherArraySeed}(x)
}
\description{
Implements a DelayedArray backend for reading from feather files.
}
\keyword{internal}
